requirements.txt
# Core Framework
streamlit>=1.31.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-google-genai>=0.0.6
langchain-groq>=0.1.9

# LLM Providers
google-generativeai>=0.3.0
groq>=0.4.0

# Financial Data
yfinance==0.2.48
requests>=2.31.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
pandas>=2.0.0

# Additional dependencies
httpx>=0.26.0
langchain-core>=0.1.0
